+++ pkg/runtime/value.go
@@ -4,6 +4,7 @@ package runtime
 import (
 	"fmt"
 	"math"
+	"sort"
 	"strconv"
 
 	"github.com/danielspk/tatu-lang/pkg/ast"
@@ -29,6 +30,7 @@ const (
 type Value interface {
 	Type() ValueType
 	String() string
+	Equal(other Value) bool
 }
 
 // Number represents a value of a number type.
@@ -57,10 +59,16 @@ func (n Number) String() string {
 		return fmt.Sprintf("%.0f", n.Value)
 	}
 
-	formatted := fmt.Sprintf("%.10f", n.Value)
-	value, _ := strconv.ParseFloat(formatted, 64)
+	return strconv.FormatFloat(n.Value, 'g', -1, 64)
+}
 
-	return fmt.Sprintf("%g", value)
+// Equal compares the number value to another.
+func (n Number) Equal(other Value) bool {
+	if other.Type() != NumberType {
+		return false
+	}
+
+	return n.Value == other.(Number).Value
 }
 
 // String represents a value of a string type.
@@ -83,6 +91,15 @@ func (s String) String() string {
 	return s.Value
 }
 
+// Equal compares the string value to another.
+func (s String) Equal(other Value) bool {
+	if other.Type() != StringType {
+		return false
+	}
+
+	return s.Value == other.(String).Value
+}
+
 // Bool represents a value of a boolean type.
 type Bool struct {
 	Value bool
@@ -103,6 +120,15 @@ func (b Bool) String() string {
 	return fmt.Sprintf("%v", b.Value)
 }
 
+// Equal compares the boolean value to another.
+func (b Bool) Equal(other Value) bool {
+	if other.Type() != BoolType {
+		return false
+	}
+
+	return b.Value == other.(Bool).Value
+}
+
 // Nil represents a value of nil type.
 type Nil struct{}
 
@@ -121,23 +147,28 @@ func (n Nil) String() string {
 	return "<nil>"
 }
 
+// Equal compares the nil value to another.
+func (n Nil) Equal(other Value) bool {
+	return other.Type() == NilType
+}
+
 // Vector represents a value of a vector type.
 type Vector struct {
 	Elements []Value
 }
 
 // NewVector builds a new Vector.
-func NewVector(elements []Value) Vector {
-	return Vector{elements}
+func NewVector(elements []Value) *Vector {
+	return &Vector{elements}
 }
 
 // Type returns the type of the vector value.
-func (v Vector) Type() ValueType {
+func (v *Vector) Type() ValueType {
 	return VectorType
 }
 
 // String returns the string representation of the vector value.
-func (v Vector) String() string {
+func (v *Vector) String() string {
 	out := "("
 
 	for i, e := range v.Elements {
@@ -150,6 +181,27 @@ func (v Vector) String() string {
 	out += ")"
 
 	return out
+}
+
+// Equal compares the vector value to another.
+func (v *Vector) Equal(other Value) bool {
+	if other.Type() != VectorType {
+		return false
+	}
+
+	o := other.(*Vector)
+
+	if len(v.Elements) != len(o.Elements) {
+		return false
+	}
+
+	for i, e := range v.Elements {
+		if !e.Equal(o.Elements[i]) {
+			return false
+		}
+	}
+
+	return true
 }
 
 // Map represents a value of map type.
@@ -167,22 +219,51 @@ func (m Map) Type() ValueType {
 	return MapType
 }
 
-// String returns the string representation of the map value.
+// String returns the string representation of the map value with sorted keys.
 func (m Map) String() string {
+	keys := make([]string, 0, len(m.Elements))
+
+	for k := range m.Elements {
+		keys = append(keys, k)
+	}
+
+	sort.Strings(keys)
+
 	out := "["
 
-	i := 0
-	for k, v := range m.Elements {
+	for i, k := range keys {
 		if i > 0 {
 			out += " "
 		}
-		out += fmt.Sprintf("%s %s", k, v.String())
-		i++
+
+		out += fmt.Sprintf("%s %s", k, m.Elements[k].String())
 	}
 
 	out += "]"
 
 	return out
+}
+
+// Equal compares the map value to another.
+func (m Map) Equal(other Value) bool {
+	if other.Type() != MapType {
+		return false
+	}
+
+	o := other.(Map)
+
+	if len(m.Elements) != len(o.Elements) {
+		return false
+	}
+
+	for k, mv := range m.Elements {
+		ov, ok := o.Elements[k]
+		if !ok || !mv.Equal(ov) {
+			return false
+		}
+	}
+
+	return true
 }
 
 // Function represents a user-defined function value (lambda/closure).
@@ -208,6 +289,11 @@ func (uf Function) String() string {
 	return "Function()"
 }
 
+// Equal compares the function value to another.
+func (uf Function) Equal(_ Value) bool {
+	return false
+}
+
 // NativeFunction represents a native function value.
 type NativeFunction struct {
 	Value func(args ...Value) (Value, error)
@@ -226,6 +312,11 @@ func (f NativeFunction) Type() ValueType {
 // String returns the string representation of the native function value.
 func (f NativeFunction) String() string {
 	return "NativeFunction()"
+}
+
+// Equal compares the native function value to another.
+func (f NativeFunction) Equal(_ Value) bool {
+	return false
 }
 
 // RecurBindings represents a tail-call marker value for TCO.
@@ -248,4 +339,9 @@ func (r RecurBindings) Type() ValueType {
 // String returns the string representation of the recur binding value.
 func (r RecurBindings) String() string {
 	return "__recur__"
+}
+
+// Equal compares the recur binding value to another.
+func (r RecurBindings) Equal(_ Value) bool {
+	return false
 }